requirements.txt
alembic==1.7.5
anyio==3.5.0
asgiref==3.5.0
autopep8==1.6.0
click==8.0.3
ecdsa==0.17.0
fastapi==0.73.0
fr==3.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
importlib-metadata==4.10.1
importlib-resources==5.4.0
Mako==1.1.6
MarkupSafe==2.0.1
passlib==1.7.4
psycopg2==2.9.3
pyasn1==0.4.8
pycodestyle==2.8.0
pydantic==1.9.0
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.31
starlette==0.17.1
toml==0.10.2
typing-extensions==4.0.1
uvicorn==0.17.1
zipp==3.7.0
